RC 0.1.2  - rev 1
-add User can change the System's Lang
-add Contact option (by Configuration)
-English lang added for all forms & System.
-fix quick-edit issues

+++ source/AKwin32/forms/maintenance/frmUsers.cs
@@ -77,7 +77,9 @@ namespace AKwin32.forms.maintenance
         private void btnEraseUser_Click(object sender, EventArgs e)
         {
             User usr = cboxUsers.SelectedItem as User;
+            if (usr == null) return;
             repo.Remove(usr);
+            listViewSources.Items.Clear();
             LoadControlsContent();
         }
